Split worker code for watching evm withdrawals
